requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.4
django-bootstrap4==0.0.8
django-rq==1.1.0
django-toolbelt==0.0.1
et-xmlfile==1.0.1
gunicorn==19.8.1
heroku==0.1.4
idna==2.6
jdcal==1.4.1
multidict==4.5.2
openpyxl==2.6.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dateutil==1.5
pytz==2018.4
redis==2.10.6
requests==2.18.4
rq==0.11.0
slackclient==2.1.0
static3==0.7.0
urllib3==1.22
whitenoise==3.3.1
yarl==1.3.0
